+++ main/ThreadController.h
@@ -45,7 +45,7 @@ namespace directgraph {
         void bar(int_fast32_t left, int_fast32_t top, int_fast32_t right, int_fast32_t bottom);
         void fillellipse(int_fast32_t x, int_fast32_t y, uint_fast32_t xradius, uint_fast32_t yradius);
         void setfillstyle(uint_fast8_t fillStyle, uint_fast32_t color);
-        void setlinestyle(uint_fast8_t linestyle, uint_fast32_t pattern, uint_fast32_t thickness);
+        void setlinestyle(uint_fast8_t linestyle, uint_fast16_t pattern, uint_fast32_t thickness);
         void setfillpattern(const char *fillpattern, uint_fast32_t color);
         void setbgcolor(uint_fast32_t color);
         void putpixel(int_fast32_t x, int_fast32_t y, uint_fast32_t color);
